forward:  Report packet drops in send().
We do all send() calls non-blocking to avoid possible dead locks.  That
means they may fail with EWOULDBLOCK/EAGAIN in which case the packet is
dropped.  In that case we report this on the console, so that we'd
notice something is off.  Also report other send errors, rather than
panicking.